+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add options to the docker popup.
 - Browse volumes with dired.
 
 ## 0.7.0 (2018-07-02)